feat: Добавлен роутинг. Добавлены файлы заглушки для страниц.
