+++ package.json
@@ -13,7 +13,7 @@
     "react-dom": "^17.0.2",
     "react-intersection-observer": "^8.34.0",
     "react-redux": "^7.2.9",
-    "react-router-dom": "^5.3.4",
+    "react-router-dom": "^6.17.0",
     "react-scripts": "5.0.1",
     "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
