--- src/components/app/app.module.css
@@ -1,15 +1,0 @@
-.app {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  height: 100%;
-  background-color: #131316;
-  color: #f2f2f3;
-}
-
-.content {
-  display: flex;
-  column-gap: 40px;
-  max-height: 80%;
-  overflow: hidden;
-}
