+++ src/pages/loginPage.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+import AppHeader from "../components/app-header/appHeader";
+
+export default function LoginPage() {
+
+    return (
+        <AppHeader />
+    )
+}
