+++ src/pages/mainPage.module.css
@@ -1,0 +1,15 @@
+.main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    height: 100%;
+    background-color: #131316;
+    color: #f2f2f3;
+  }
+  
+  .content {
+    display: flex;
+    column-gap: 40px;
+    max-height: 80%;
+    overflow: hidden;
+  }
